pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "flask-bcrypt>=1.0.1",
    "sqlalchemy>=2.0.39",
    "wtforms>=3.2.1",
    "apscheduler>=3.11.0",
    "werkzeug>=3.1.3",
    "cryptography>=44.0.2",
    "pyinstaller>=6.12.0",
    "pyperclip>=1.9.0",
    "psutil>=7.0.0",
    "os-sys>=0.9.1",
]
